requirements.txt
python==3.6.9
pytorch==1.10.2
scipy==1.5.2
numpy==1.19.5
librosa==0.8.1
pyworld==0.3.0
tqdm==4.62.3